backend/pyproject.toml
[project]
name = "d-done-backend"
version = "0.2.0"
description = "D-Done — AI Due Diligence platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "google-adk>=1.0.0",
    "google-cloud-logging>=3.8.0",
    "google-cloud-storage>=2.19.0",
    "google-genai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "descope>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "google-adk[eval]",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
